Added function and UI to display tasks
