--- views/todo.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" >
-<head>
-  <meta charset="UTF-8">
-  <title>Tasksp</title>
-  <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/normalize/5.0.0/normalize.min.css">
-  <link rel="stylesheet" href="../static/credentials.css">
-</head>
-
-<body>
-  <p> TO DO</p>
-  
-</body>
-</html>
